requirements.txt
python-dotenv
psycopg2-binary
discord.py
riotwatcher
SQLAlchemy
alembic
pytest
logging-colorizer
pydantic
aiohttp